+++ src/main/java/edu/ie3/simona/api/data/em/ExtEmData.java
@@ -9,6 +9,7 @@ package edu.ie3.simona.api.data.em;
 import edu.ie3.datamodel.models.value.PValue;
 import edu.ie3.datamodel.models.value.Value;
 import edu.ie3.simona.api.data.ExtData;
+import edu.ie3.simona.api.data.ExtInputDataPackage;
 import edu.ie3.simona.api.data.em.ontology.EmDataMessageFromExt;
 import edu.ie3.simona.api.data.em.ontology.ProvideEmData;
 import edu.ie3.simona.api.data.ontology.ScheduleDataServiceMessage;
@@ -26,21 +27,37 @@ import java.util.UUID;
 public class ExtEmData implements ExtData {
 
   /** Actor reference to service that handles ev data within SIMONA */
-  private final ActorRef dataService;
+  private ActorRef dataService;
 
   /** Actor reference to adapter that handles scheduler control flow in SIMONA */
-  private final ActorRef extSimAdapter;
+  private ActorRef extSimAdapter;
 
   /** Assets that provide primary data to SIMONA */
-  private final List<UUID> controlledEms;
+  private final Map<String, UUID> extEmMapping;
 
-  public ExtEmData(ActorRef dataService, ActorRef extSimAdapter, EmDataFactory factory, List<UUID> controlledEms) {
-    this.dataService = dataService;
-    this.extSimAdapter = extSimAdapter;
-    this.controlledEms = controlledEms;
+  private final EmDataFactory emDataFactory;
+
+  public ExtEmData(
+          EmDataFactory emDataFactory,
+          Map<String, UUID> extEmMapping
+  ) {
+    this.emDataFactory = emDataFactory;
+    this.extEmMapping = extEmMapping;
   }
 
-  public List<UUID> getControlledEms() { return controlledEms; }
+  public void setActorRefs(
+          ActorRef dataService,
+          ActorRef extSimAdapter
+  ) {
+    this.dataService = dataService;
+    this.extSimAdapter = extSimAdapter;
+  }
+
+  public List<UUID> getControlledEms() { return extEmMapping.values().stream().toList(); }
+
+  public EmDataFactory getEmDataFactory() {
+    return emDataFactory;
+  }
 
   /** Provide primary data from an external simulation in one tick. */
   public void provideEmData(Long tick, Map<UUID, PValue> emData) {
@@ -58,5 +75,27 @@ public class ExtEmData implements ExtData {
     dataService.tell(msg, ActorRef.noSender());
     // we need to schedule data receiver activation with scheduler
     extSimAdapter.tell(new ScheduleDataServiceMessage(dataService), ActorRef.noSender());
+  }
+
+
+  public Map<UUID, PValue> createExtEmDataMap(
+          ExtInputDataPackage extEmData
+  ) {
+    Map<UUID, PValue> emDataForSimona = new HashMap<>();
+    extEmData.getSimonaInputMap().forEach(
+            (id, extInput) -> {
+              if (extEmMapping.containsKey(id)) {
+                try {
+                  emDataForSimona.put(
+                          extEmMapping.get(id),
+                          emDataFactory.convert(extInput)
+                  );
+                } catch (ConvertionException e) {
+                  throw new RuntimeException(e);
+                }
+              }
+            }
+    );
+    return emDataForSimona;
   }
 }
