+++ src/main/java/edu/ie3/simona/api/data/primarydata/ExtPrimaryData.java
@@ -8,6 +8,7 @@ package edu.ie3.simona.api.data.primarydata;
 
 import edu.ie3.datamodel.models.value.Value;
 import edu.ie3.simona.api.data.ExtData;
+import edu.ie3.simona.api.data.ExtInputDataPackage;
 import edu.ie3.simona.api.data.ontology.ScheduleDataServiceMessage;
 import edu.ie3.simona.api.data.primarydata.ontology.PrimaryDataMessageFromExt;
 import edu.ie3.simona.api.data.primarydata.ontology.ProvidePrimaryData;
@@ -21,25 +22,36 @@ import org.apache.pekko.actor.ActorRef;
 public class ExtPrimaryData implements ExtData {
 
   /** Actor reference to service that handles ev data within SIMONA */
-  private final ActorRef dataService;
+  private ActorRef dataService;
 
   /** Actor reference to adapter that handles scheduler control flow in SIMONA */
-  private final ActorRef extSimAdapter;
+  private ActorRef extSimAdapter;
 
   /** Factory to convert an external object to PSDM primary data */
-  private final PrimaryDataFactory factory;
+  private final PrimaryDataFactory primaryDataFactory;
 
   /** Assets that provide primary data to SIMONA */
-  private final List<UUID> primaryDataAssets;
+  private final Map<String, UUID> extPrimaryDataMapping;
 
-  public ExtPrimaryData(ActorRef dataService, ActorRef extSimAdapter, PrimaryDataFactory factory, List<UUID> primaryDataAssets) {
-    this.dataService = dataService;
-    this.extSimAdapter = extSimAdapter;
-    this.factory = factory;
-    this.primaryDataAssets = primaryDataAssets;
+  public ExtPrimaryData(PrimaryDataFactory primaryDataFactory, Map<String, UUID> extPrimaryDataMapping) {
+    this.primaryDataFactory = primaryDataFactory;
+    this.extPrimaryDataMapping = extPrimaryDataMapping;
   }
 
-  public List<UUID> getPrimaryDataAssets() { return primaryDataAssets; }
+  public void setActorRefs(
+          ActorRef dataService, ActorRef extSimAdapter
+  ) {
+    this.dataService = dataService;
+    this.extSimAdapter = extSimAdapter;
+  }
+
+  public PrimaryDataFactory getPrimaryDataFactory() {
+    return primaryDataFactory;
+  }
+
+  public List<UUID> getPrimaryDataAssets() {
+    return extPrimaryDataMapping.values().stream().toList();
+  }
 
   /** Provide primary data from an external simulation in one tick. */
   public void providePrimaryData(Long tick, Map<UUID, Value> primaryData) {
@@ -57,5 +69,26 @@ public class ExtPrimaryData implements ExtData {
     dataService.tell(msg, ActorRef.noSender());
     // we need to schedule data receiver activation with scheduler
     extSimAdapter.tell(new ScheduleDataServiceMessage(dataService), ActorRef.noSender());
+  }
+
+  public Map<UUID, Value> createExtPrimaryDataMap(
+          ExtInputDataPackage extInputDataPackage
+  ) {
+    Map<UUID, Value> primaryDataForSimona = new HashMap<>();
+    extInputDataPackage.getSimonaInputMap().forEach(
+            (id, extInput) -> {
+              if (extPrimaryDataMapping.containsKey(id)) {
+                try {
+                  primaryDataForSimona.put(
+                          extPrimaryDataMapping.get(id),
+                          primaryDataFactory.convert(extInput)
+                  );
+                } catch (ConvertionException e) {
+                  throw new RuntimeException(e);
+                }
+              }
+            }
+    );
+    return primaryDataForSimona;
   }
 }
