+++ src/main/java/edu/ie3/simona/api/simulation/ExtSimulation.java
@@ -140,4 +140,6 @@ public abstract class ExtSimulation implements Runnable {
   protected String[] getMainArgs() {
     return data.getMainArgs();
   }
+
+  public abstract List<ExtData> getDataConnections();
 }
